broke out description audio into its own component since it's podcast transcript, not description and super long
